update with latest node template features
