+++ web/frontend/.gitignore
@@ -16,6 +16,3 @@ dist/
 # Without this, they may get a warning if using a different package manager to us
 yarn.lock
 package-lock.json
-
-# Unneccessary File
-commented.jsx
