--- web/frontend/components/OrderDetails.jsx
@@ -1,8 +1,0 @@
-import React from 'react'
-
-export function OrderDetails() {
-  return (
-    <div>OrderDetails</div>
-  )
-}
-
