--- web/frontend/components/OrderGraphs.jsx
@@ -1,8 +1,0 @@
-import React from 'react'
-
-
-export function OrderGraphs() {
-    return (
-        <div>OrderGraphs</div>
-    )
-}
