+++ web/frontend/components/providers/index.js
@@ -1,3 +1,2 @@
 export { QueryProvider } from "./QueryProvider";
 export { PolarisProvider } from "./PolarisProvider";
-export {AppBridgeProvider} from "./AppBridgeProvider";
