--- web/frontend/data.js
@@ -1,6 +1,0 @@
-export const storeData = [
-    { name: "Matcha Store", year: 1985, order: 20, completed: 10 },
-    { name: "Burger Store", year: 1988, order: 12, completed: 6 },
-    { name: "Street Store", year: 1990, order: 15, completed: 5 },
-    { name: "Flower Store", year: 1995, order: 18, completed: 8 },
-];
