--- web/frontend/hooks/index.js
@@ -1,2 +1,0 @@
-export { useAppQuery } from "./useAppQuery";
-export { useAuthenticatedFetch } from "./useAuthenticatedFetch";
