+++ web/frontend/index.html
@@ -17,7 +17,7 @@
       }
     </script>
 
-    <meta name="shopify-api-key" content="48a0f5e2ec6e42fae077af3ea19bb608" />
+    <meta name="shopify-api-key" content="%VITE_SHOPIFY_API_KEY%" />
     <script src="https://cdn.shopify.com/shopifycloud/app-bridge.js"></script>
   </head>
   <body>
