+++ web/frontend/index.jsx
@@ -1,7 +1,6 @@
 import App from "./App";
 import { createRoot } from "react-dom/client";
 import { initI18n } from "./utils/i18nUtils";
-import "./App.css"
 
 // Ensure that locales are loaded before rendering the app
 initI18n().then(() => {
