+++ web/frontend/package.json
@@ -21,16 +21,13 @@
     "@formatjs/intl-locale": "^3.3.2",
     "@formatjs/intl-localematcher": "^0.4.0",
     "@formatjs/intl-pluralrules": "^5.2.4",
-    "@mui/icons-material": "^6.3.1",
-    "@mui/material": "^6.3.1",
-    "@mui/styled-engine-sc": "^6.3.1",
+    "@mui/icons-material": "^6.4.0",
+    "@mui/material": "^6.4.0",
     "@shopify/app-bridge": "^3.7.7",
-    "@shopify/app-bridge-react": "^3.7.7",
+    "@shopify/app-bridge-react": "^4.1.3",
     "@shopify/i18next-shopify": "^0.2.9",
-    "@shopify/polaris": "^10.50.1",
-    "@shopify/polaris-icons": "^9.3.0",
+    "@shopify/polaris": "^10.49.1",
     "@vitejs/plugin-react": "4.2.1",
-    "dotenv": "^16.4.7",
     "i18next": "^23.1.0",
     "i18next-resources-to-backend": "^1.1.4",
     "react": "^18.2.0",
@@ -38,7 +35,6 @@
     "react-i18next": "^13.0.0",
     "react-query": "^3.34.19",
     "react-router-dom": "^6.3.0",
-    "styled-components": "^6.1.14",
     "vite": "^5.4.0"
   },
   "devDependencies": {
