+++ web/package.json
@@ -13,16 +13,12 @@
     "node": ">=16.13.0"
   },
   "dependencies": {
-    "@shopify/shopify-api": "^11.6.1",
-    "@shopify/shopify-app-express": "^5.0.9",
+    "@shopify/shopify-app-express": "^5.0.8",
     "@shopify/shopify-app-session-storage-sqlite": "^4.0.8",
-    "axios": "^1.7.9",
     "compression": "^1.7.4",
-    "cookie-parser": "^1.4.7",
     "cross-env": "^7.0.3",
     "dotenv": "^16.4.7",
     "pg": "^8.13.1",
-    "pg-hstore": "^2.3.4",
     "sequelize": "^6.37.5",
     "serve-static": "^1.14.1"
   },
